moved stylings to new folder
