+++ src/assets/loaderStyles.jsx
@@ -1,0 +1,14 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles(theme => ({
+  root: {
+    marginLeft: '50%',
+    marginTop: '150px',
+    display: 'flex',
+    '& > * + *': {
+      marginLeft: theme.spacing(2),
+    },
+  },
+}));
+
+export default useStyles;
