+++ src/elements/loader.jsx
@@ -1,17 +1,6 @@
 import React from 'react';
-import { makeStyles } from '@material-ui/core/styles';
 import CircularProgress from '@material-ui/core/CircularProgress';
-
-const useStyles = makeStyles(theme => ({
-  root: {
-    marginLeft: '50%',
-    marginTop: '150px',
-    display: 'flex',
-    '& > * + *': {
-      marginLeft: theme.spacing(2),
-    },
-  },
-}));
+import useStyles from '../assets/loaderStyles';
 
 export default function Loader() {
   const classes = useStyles();
