+++ src/elements/mainAvatar.jsx
@@ -1,18 +1,8 @@
 import React from 'react';
-import { makeStyles } from '@material-ui/core/styles';
 import Grid from '@material-ui/core/Grid';
 import Avatar from '@material-ui/core/Avatar';
 import Box from '@material-ui/core/Box';
-
-const useStyles = makeStyles(theme => ({
-  large: {
-    width: theme.spacing(20),
-    height: theme.spacing(20),
-  },
-  bgColor: {
-    backgroundColor: 'rgba(255, 255, 255, 0.8)',
-  },
-}));
+import useStyles from '../assets/mainAvatarStyles';
 
 export default function MainAvatar() {
   const classes = useStyles();
